kafka_rw/requirements.txt
isort==5.10.1
kafka-python==2.0.2
pydantic==1.9.0
python-json-logger==2.0.2